pyproject.toml
[project]
name = "fastvoicechat"
version = "0.0.0"
description = "高速音声対話のデモプログラム"
readme = "README.md"
authors = [
    { name = "shimajiroxyz" }
]
requires-python = ">=3.11"
dependencies = [
    "google-cloud-speech>=2.31.0",
    "openai>=1.64.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "setuptools>=75.8.0",
    "simpleaudio>=1.0.4",
    "webrtcvad>=2.0.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
